integrate in-app-notification interfaces
